Rename sensor.h and sensor_private.h
