+++ include/platform/sensor.h
