+++ include/platform/sensor_private.h
